Move calculateAge to shared folder
