+++ src/applications/dependents/dependents-verification/helpers.js
@@ -1,68 +1,9 @@
 import React from 'react';
-import {
-  parse,
-  isValid,
-  differenceInYears,
-  differenceInMonths,
-  differenceInDays,
-  format,
-} from 'date-fns';
+import { parse, format } from 'date-fns';
+
+import { calculateAge } from '../shared/utils';
 
 export const hasSession = () => localStorage.getItem('hasSession') === 'true';
-
-export const calculateAge = dob => {
-  if (!dob) {
-    return {
-      age: 0,
-      dobStr: '',
-      labeledAge: '',
-    };
-  }
-
-  const dobObj = parse(dob, 'MM/dd/yyyy', new Date());
-  const dobStr = isValid(dobObj) ? format(dobObj, 'MMMM d, yyyy') : '';
-  const invalid = {
-    age: 0,
-    dobStr,
-    labeledAge: '',
-  };
-
-  if (!dobStr) {
-    return invalid;
-  }
-
-  const ageInYears = differenceInYears(new Date(), dobObj);
-  if (ageInYears > 0) {
-    return {
-      age: ageInYears,
-      dobStr,
-      labeledAge: `${ageInYears} year${ageInYears > 1 ? 's' : ''} old`,
-    };
-  }
-
-  const ageInMonths =
-    ageInYears > 0 ? 0 : differenceInMonths(new Date(), dobObj);
-  if (ageInMonths > 0) {
-    return {
-      age: 0,
-      dobStr,
-      labeledAge: `${ageInMonths} month${ageInMonths > 1 ? 's' : ''} old`,
-    };
-  }
-
-  // If less than a month old, show days
-  const ageInDays = ageInMonths > 0 ? 0 : differenceInDays(new Date(), dobObj);
-  return ageInDays >= 0
-    ? {
-        age: 0,
-        dobStr,
-        labeledAge:
-          ageInDays === 0
-            ? 'Newborn'
-            : `${ageInDays} day${ageInDays > 1 ? 's' : ''} old`,
-      }
-    : invalid;
-};
 
 /**
  * @typedef DependentsArrayFromAPI
