+++ src/applications/dependents/dependents-verification/tests/unit/components/DependentsInformation.unit.spec.jsx
@@ -8,7 +8,7 @@ import { $, $$ } from 'platform/forms-system/src/js/utilities/ui';
 
 import { DependentsInformation } from '../../../components/DependentsInformation';
 import { defaultData } from './dependent-data';
-import { calculateAge } from '../../../helpers';
+import { calculateAge } from '../../../../shared/utils';
 
 function renderPage({
   data = defaultData,
