+++ src/applications/dependents/dependents-verification/tests/unit/helpers.unit.spec.jsx
@@ -1,7 +1,6 @@
 import { expect } from 'chai';
-import { sub, format } from 'date-fns';
 
-import { hasSession, calculateAge, processDependents } from '../../helpers';
+import { hasSession, processDependents } from '../../helpers';
 
 describe('hasSession', () => {
   it('should return true if hasSession is set to true in localStorage', () => {
@@ -13,125 +12,6 @@ describe('hasSession', () => {
     expect(hasSession()).to.be.false;
   });
 });
-
-describe('calculateAge', () => {
-  it('should return 0 age and empty strings if dob is not provided', () => {
-    const result = calculateAge();
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: '',
-      labeledAge: '',
-    });
-  });
-
-  it('should return 0 age and empty strings if dob is invalid', () => {
-    const result = calculateAge('');
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: '',
-      labeledAge: '',
-    });
-  });
-
-  it('should return 0 age and correct dobStr for future dates', () => {
-    const testDate = sub(new Date(), {
-      months: -1,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '',
-    });
-  });
-
-  it('should correctly calculate age an 18 year old', () => {
-    const testDate = sub(new Date(), {
-      years: 18,
-      months: 1,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 18,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '18 years old',
-    });
-  });
-
-  it('should correctly calculate age for a 1 year old', () => {
-    const testDate = sub(new Date(), {
-      years: 1,
-      days: 3,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 1,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '1 year old',
-    });
-  });
-
-  it('should correctly calculate age for a 4 month old', () => {
-    const testDate = sub(new Date(), {
-      months: 4,
-      days: 3,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '4 months old',
-    });
-  });
-
-  it('should correctly calculate age for a 1 month old', () => {
-    const testDate = sub(new Date(), {
-      months: 1,
-      days: 1,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '1 month old',
-    });
-  });
-
-  it('should correctly calculate age for a 10 day old', () => {
-    const testDate = sub(new Date(), {
-      days: 10,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '10 days old',
-    });
-  });
-
-  it('should correctly calculate age for a 1 day old', () => {
-    const testDate = sub(new Date(), {
-      days: 1,
-    });
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: '1 day old',
-    });
-  });
-
-  it('should correctly calculate age for a baby born today', () => {
-    const testDate = new Date();
-    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
-    expect(result).to.deep.equal({
-      age: 0,
-      dobStr: format(testDate, 'MMMM d, yyyy'),
-      labeledAge: 'Newborn',
-    });
-  });
-});
-
 describe('processDependents', () => {
   it('should return an empty array when no dependents are provided', () => {
     expect(processDependents()).to.deep.equal([]);
