+++ src/applications/dependents/shared/tests/utils/utils.unit.spec.jsx
@@ -1,5 +1,6 @@
 import { expect } from 'chai';
 import { render } from '@testing-library/react';
+import { format, sub } from 'date-fns';
 
 import {
   getFullName,
@@ -7,6 +8,7 @@ import {
   maskID,
   isEmptyObject,
   getRootParentUrl,
+  calculateAge,
 } from '../../utils';
 
 describe('getFullName', () => {
@@ -128,5 +130,123 @@ describe('getFormatedDate', () => {
   it('should return formatted date for valid Date object', () => {
     const formattedDate = getFormatedDate('12/05/2022', 'MM/dd/yyyy', 'PPPP');
     expect(formattedDate).to.equal('Monday, December 5th, 2022');
+  });
+});
+
+describe('calculateAge', () => {
+  it('should return 0 age and empty strings if dob is not provided', () => {
+    const result = calculateAge();
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: '',
+      labeledAge: '',
+    });
+  });
+
+  it('should return 0 age and empty strings if dob is invalid', () => {
+    const result = calculateAge('');
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: '',
+      labeledAge: '',
+    });
+  });
+
+  it('should return 0 age and correct dobStr for future dates', () => {
+    const testDate = sub(new Date(), {
+      months: -1,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: 'Date in the future',
+    });
+  });
+
+  it('should correctly calculate age an 18 year old', () => {
+    const testDate = sub(new Date(), {
+      years: 18,
+      months: 1,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 18,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: '18 years old',
+    });
+  });
+
+  it('should correctly calculate age for a 1 year old', () => {
+    const testDate = sub(new Date(), {
+      years: 1,
+      days: 3,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 1,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: '1 year old',
+    });
+  });
+
+  it('should correctly calculate age for a 4 month old', () => {
+    const testDate = sub(new Date(), {
+      months: 4,
+      days: 3,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: '4 months old',
+    });
+  });
+
+  it('should correctly calculate age for a 1 month old', () => {
+    const testDate = sub(new Date(), {
+      months: 1,
+      days: 1,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: '1 month old',
+    });
+  });
+
+  it('should correctly calculate age for a 10 day old', () => {
+    const testDate = sub(new Date(), {
+      days: 10,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: '10 days old',
+    });
+  });
+
+  it('should correctly calculate age for a 1 day old', () => {
+    const testDate = sub(new Date(), {
+      days: 1,
+    });
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: '1 day old',
+    });
+  });
+
+  it('should correctly calculate age for a baby born today', () => {
+    const testDate = new Date();
+    const result = calculateAge(format(testDate, 'MM/dd/yyyy'));
+    expect(result).to.deep.equal({
+      age: 0,
+      dobStr: format(testDate, 'MMMM d, yyyy'),
+      labeledAge: 'Newborn',
+    });
   });
 });
